requirements.txt
# Core Deep Learning
tensorflow>=2.10.0
keras>=2.10.0

# YOLO Object Detection
ultralytics>=8.0.0

# Image Processing
opencv-python>=4.7.0
Pillow>=9.0.0

# Data Manipulation
numpy>=1.23.0
pandas>=1.5.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# Jupyter Notebook
jupyter>=1.0.0
ipykernel>=6.0.0

# Utilities
scikit-learn>=1.2.0
tqdm>=4.64.0
pyyaml>=6.0

# Dataset handling
requests>=2.28.0
gdown>=4.6.0
